fix form send data
